Create generateHintMa.c
Create a function, generateHintMap,
that accepts an array of zeroes (0) and
ones (1),field , where 1 represent the
mines for the Minesweeper game and 0 is
a safe space. The function must return an
array of integers containing the number of
mines around a specific location
(including diagonal). If the location
contains a mine, assign it with -1.